requirements.txt
llama-index==0.8.56
openai==0.28.1
jsonpath_ng==1.6.0
python-dotenv==1.0.0
